Emphasize that the Subject.check overload that users should use accepts parameters.
This also fixes some Javadoc links that the Javadoc tool currently resolves to the package-private overload.

And remove some TODOs now that Subject subclassing works better (and we're stuck with it, anyway, because of 1.0 :))

Finally, link to a bug that covers a TODO (among other things it covers).

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=255660230