Add a reflective field equality based subject and correspondence.
This is written in Kotlin and depends on kotlin runtime as a result, however is designed to work cleanly from Java, and its kotlin origins should not be paritcularly noticeable to the test-writing developer.

The tests handle single object testing (via the Subject) and list element comparisons using a Correspondence. Field comparisons can be shallow or deep, and can respect (in the deep case) declared `equals()` methods or ignore them.  It also can handle reference cycles (with caveats).

This is upstreamed from Square, Inc. and licensed under Apache 2.0.