Slight tweak to make IterableSubject.UsingCorrespondence.contains() failures more like IterableSubject.contains().
The non-fuzzy version only includes a "but did not" fact and replaces the "but was" fact with a "full contents" fact when there is a "though it did contain" fact. This change makes the fuzzy version behave the same way.

Note that it still says "full contents" in the "found match (but failing because of exception)" case, since "but was" seems wrong for this case.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=257852478