Fix null key bug in ProtoTruthMessageDifferencer#toProtoMap
Maps are implemented in protobufs as repeated messages with a 'key' and 'value' field. In Proto3 fields whose values equal the defaultValue for the field are not serialised to the wire format. That means when the bytes are parsed to de-serialise the message this field is absent, and calls to get(FieldDescriptor) for that field will return null.

This should be interpreted as the defaultValue. Maps in generated message classes already handle this case but DynamicMessage does not, so it must be explicitly handled here.

Also this CL changes the return type of toProtoMap to be an ImmutableMap, where previously it returned either mutable or immutable implementations of Map depending on if the supplied container was null.

RELNOTES=Fix handling of Map with keys equal to the key types default value.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=281482928