Get rid of the Facts class.
I think that this class was an interesting experiment which, but that in the end it didn't quite justify its existence. It's another concept that the reader needed to know about, and the boilerplate resulting from just using a list isn't too bad, and everyone knows how lists work.

This CL suggests some opportunities for commoning up code between branches of conditionals and the like here. I'm not going to do that in this CL, since the next CLs are going to make changes to the Facts we produce, to give more structured information.

I also do a couple of bits of cleanup, mostly suggested by ErrorProne, and suggest a couple of warnings.

RELNOTES=n/a

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=314724975