Clarify creation date of AssertJ in comparison.md, too.
Also, rephrase a claim that probably gives the wrong impression: Google's usage of Truth at the time AssertJ was _created_ was in the hundreds of files. While that's arguably "widely" (e.g., it crossed many teams), it's not enough to be a major _barrier to migration_ the way that our usage became in the next year or so.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=324894100