Add a mention about proto-lite support early in the doc
When looking for Truth support for proto-lite it is easy to
see the first part of the doc, try it out, realize it is not
the right instructions, then find the proto-lite support details.
This CL adds a mention early on to help proto-lite users find the
appropriate instructions at the bottom of the doc before trying
the proto instructions.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=325887655