Replace the unicode quote markers (“) with ascii quote markers (")
When the page is rendered for the external site the quote markers
are displayed in red, as if they were an error. Replacing them with
ascii quotes should highlight the string and remove the red.
Also, if the example were copy/pasted it would work instead of
resulting in a compile failure.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=325892218