Use the same version for protoc as we do for the protobuf runtime.
When we were running our open-source tests on our internal presubmit system, we had problems because the newer protoc binaries were built against a different runtime that our presubmit system used. Now that we're running our open-source tests in external CI, this is no longer an issue.

While we're at it, I was also going to also use a newer protoc for our Lite support. From a quick glance, I didn't see any reason that we needed a different version there. I had guessed that we just didn't standardize the version when we moved the version properties from the child poms to the parent pom. However, when I try that, I see that newer versions of protoc don't support the way we currently request a Lite runtime. Rather than try to figure out the right way to request the Lite runtime now, I'm keeping the old protoc for our Lite support for now.

My hope is that, since this removes the scrubbing directives from the top-level pom.xml, we'll now be able to use Copybara to import PRs to update version numbers in that file.

RELNOTES=n/a
PiperOrigin-RevId: 335876304