Bump Truth to Guava 30.0, and make the version in our POM more automation-friendly.
Currently, Dependabot can't bump the Guava version for us, since our version property holds "30.0," rather than "30.0-jre" or "30.0-android."

To fix that, this CL switches us to having 2 separate version properties, each of which holds a full version number. This lets Dependabot (and regular Maven tools) bump them automatically in the future. It does mean that we'll want to keep the 2 properties in sync, so we'll see if Dependabot is smart enough to update both in the same PR in the future.

Also: To avoid a dependencyConvergence violation, we need to specify a specific version for another transitive dependency, jsinterop-annotations. (As already noted in the comments, we should possibly just eliminate our dependencyConvergence check entirely. Also, we should possibly eliminate _dependencyManagement_ entirely in favor of just declaring regular dependencies on everything. Doing so might help us work around a requireUpperBounds bug -- and requireUpperBounds is the important thing.)
RELNOTES=n/a
PiperOrigin-RevId: 338058375