When falling back to plain AssertionError, still include the "expected" and "but was" lines.
I have no idea what I was thinking when I went out of my way to omit them in CL 353141904.

And I think it had crossed my mind at one point to test the contents of the AssertionError, but I did not....

This CL does *not* address the main discussion of the day from https://github.com/google/truth/issues/333#issuecomment-765616282; it's just *another* bug I noticed along the way. Sigh.

RELNOTES=When JUnit 4 is excluded from the classpath, the `AssertionError` Truth generates as a substitute for `ComparisonFailure` now includes the expected and actual values that were missing in 1.1.1.
PiperOrigin-RevId: 353319232