Write a little more about _why_ to avoid helper methods.
As part of that, slightly expand the scope of this section to talk about helper methods in general, rather than just those that _return_ a `Subject` instance.

Arguably I should more significantly rewrite and restructure this section -- "Prefer not to wrap `assertThat` calls in helper methods" or something, with the existing text about _returning_ `Subject` instances as a particular example of that? But hopefully this is good enough for now.

I took some of the content from CL 134823744. Other parts came to mind because of the recent discussions on b/176092991#comment2 and http://b/175894396#comment2.

RELNOTES=n/a
PiperOrigin-RevId: 355172125