Hopefully fix issue with snapshot deployment failing.
Based on the error for https://github.com/google/truth/runs/2122895842?check_suite_focus=true, something (Maven?) is passing `--no-module-directories` to `javadoc` despite that we're trying to run this on JDK 8 which didn't have that flag. We could perhaps skip Javadoc generation for snapshot deployment (we were before), but I feel like it's better to include Javadoc if we can.

RELNOTES=n/a
PiperOrigin-RevId: 363190808