Remove some `@Nullable` annotations from type-variable usages.
The way that these particular parameters work is that they are "nullable if the type argument is nullable, and non-null if the type argument is non-null." In modern ways of annotating for nullness, this is expressed by writing plain `T`, rather than `@Nullable T`.

Historically, we've sometimes had to write `@Nullable T` to counteract the effects of `@ParametersAreNonnullByDefault`, particularly under Kotlin. But we don't use `@ParametersAreNonnullByDefault` here, so it's not an issue.

(Sometimes we've also had to write it to satisfy `NullPointerTester`. But we mostly don't use that for Truth, either.)

RELNOTES=n/a
PiperOrigin-RevId: 389861654