Annotate the constructor of `ComparisonFailureWithFacts` with `@UsedByReflection`.
This makes it possible to write a config that instructs Proguard to keep that constructor if it keeps the class.

I'm hoping that I don't need to explicitly instruct it to keep the class, since the class name appears as a literal argument to `Class.forName`. (I could just tell it to keep it "just to be safe," but then maybe we'd see problems with users who don't actually use Truth but end up with it at runtime anyway because it's somewhere in the transitive deps and we asked Proguard to keep it.)

(The reason we're using reflection in the first place is to support running without JUnit 4 on the classpath. See https://github.com/google/truth/issues/333.)

RELNOTES=n/a
PiperOrigin-RevId: 462826082