Use `@Nullable` instead of `@CheckForNull`.
We'd been using the latter to work around a bug in Kotlin, but that bug was fixed in 1.8.20, and Google has on that version (or later) for a while.

(I think the bug is the flip side of [KT-57996](https://youtrack.jetbrains.com/issue/KT-57996/Usages-of-Foo-Nullable-produce-only-warnings-even-with-Xtype-enhancement-improvements-strict-mode-Xjspecify-annotationsstrict): In KT-57996, Kotlin didn't know that an annotated array was nullable, so it didn't produce the expected warning/error until 1.8.20. Here, Kotlin didn't know that an annotated array was nullable, so it prevented callers from passing null until 1.8.20.)

If Truth users outside Google are still using 1.8.20, they might be affected; I'm not completely sure. It might depend on whether they set `-Xtype-enhancement-improvements-strict-mode`??

RELNOTES=n/a
PiperOrigin-RevId: 532397177