Support deep comparison of unpacked Any messages in FieldNumberTree.
RELNOTES=Fixed a bug that caused ProtoTruth to ignore the contents of unpacked `Any` messages. This fix may cause tests to fail, since ProtoTruth will now check whether the message contents match. If so, you may need to change the values that your tests expect, or there may be a bug in the code under test that had been hidden by the Truth bug. Sorry for the trouble.
PiperOrigin-RevId: 577171522