Remove a confusing comment.
When I saw a Copybara update for this dependency I wondered whether the comment meant that the version had to stay at a particular level to "work around a ubuntu bug". But I think it was just saying, in 2015, that the default maven-jar-plugin version had the bug and we needed to use a more recent explicit version.

RELNOTES=n/a
PiperOrigin-RevId: 624940222