Explain how `isSameInstanceAs` treats nulls.
We could do better here in explaining that. We could also do better in various other ways, such as rewriting our messages in a way that doesn't begin "Fails if...." (I think we used to have "Attests that...," which I was also soured on. I wonder if I switched us to "Fails if..." at that point :)) The one additional thing I did do was change the text (in this one particular instance) to stop [referring to the actual value under test as "the subject,"](https://github.com/google/truth/issues/203) since that's ambiguous between the `Subject` instance and the value that it holds.

RELNOTES=n/a
PiperOrigin-RevId: 680668770