Update comments about why `ActualValueInference` might not load.
`ActualValueInference`+ASM has been the default for Android builds inside Google (cl/492596588) and for external builds (cl/336875929) for a while now. It's still possible for users to manually exclude ASM or to run an optimizer that removes `ActualValueInference` (and ASM along with it). This CL attempts to make that clearer.

RELNOTES=n/a
PiperOrigin-RevId: 693060958