Use `@code` instead of `@link` for primitive types.
`@link` is wrong, and accordingly, it produces Javadoc warnings. Those warnings then made it hard for me to see the actual error I was seeing, which was a result of trying to use Java 21's version of Javadoc. I'll leave a note about that on cl/647714634.

(followup to cl/681080215)

RELNOTES=n/a
PiperOrigin-RevId: 696692285