Decouple the version of GWT from the version of the GWT Maven Plugin, and bump GWT.
The plugin that we use has stopped being upgraded. We should really migrate off it, but in the meantime, we can at least upgrade GWT.

It turns out that we can't upgrade to the _newest_ version of GWT because its version of the Eclipse compiler detects a mismatch between the `guava-android` and `guava-jre` _internal implementation code_, since the Truth tests mix the two. I guess that Dependabot will try to upgrade us, anyway. We can [try](https://github.com/google/truth/pull/1348) to tell it to stop.

RELNOTES=n/a
PiperOrigin-RevId: 711507563