Standardize Dependabot configs on "Maven weekly, GitHub Actions monthly."
This includes:
- setting up Dependabot _at all_ for a few projects
- dropping GitHub Actions from weekly to monthly for the rest

My feeling on the latter is that GitHub Actions upgrades never feel urgent: Even when GitHub stopped supporting old versions of `actions/cache`, they gave plenty of warning. I'd also note that I don't think we've had trouble much (if ever?) with upgrades to GitHub Actions, so there's even less reason to fear batching of updates than usual. Given that, we might as well try to batch together as many updates as we can so as to marginally reduce toil. (And if an upgrade it ever truly urgent for security reasons, I expect that Dependabot would push us to it promptly, anyway, perhaps even for projects without a Dependabot config at all.)

RELNOTES=n/a
PiperOrigin-RevId: 739294702