Say "Checks that _x_," not "Fails if _!x_."
We've moved a bit in this direction over the years, and now I'm trying to push us the rest of the way.

I did most of this with regex search and replace, so it's likely that I've introduced the sorts of errors you might expect from that (in addition to whatever errors I have introduced by hand :)). To aid the automation, I did end up stamping out phrases like "Checks that _the string_...," as previously discussed in cl/766689205... only to reintroduce them for _some_ types. I hope that we are at least now largely consistent within a given type.

Just to make things interesting, I've tweaked phrasing in a few other ways. The most notable change is to back out some discussion of duplicates from methods like `containsNoneOf`. That discussion was introduced in cl/130379377, so of course push back if it seems worth keeping.

RELNOTES=n/a
PiperOrigin-RevId: 766810480