[Fail instead of throwing NPE](https://github.com/google/truth/issues/207) for `assertThat(nullClass).isAssignableTo(...)`.
I could see an argument that we might want to be more explicit here that the user passed a _class_ that is null, rather than passing a null _object_ (since null can naturally be considered to be assignable to any type). Let me know what you think.

RELNOTES=n/a
PiperOrigin-RevId: 769696908