Improve some numeric failure cases:
- [Fail instead of throwing an exception](https://github.com/google/truth/issues/207) for null actual values and invalid tolerances.
- Clarify failure messages around "weird" actual and expected values.

Then touch up `MathUtil` and other classes while in the area.

Plus, remove an `actual` local variable in the `isWithin`+`isNotWithin` lambdas that seems to have been a holdover from when we had to write "`getSubject()`" instead of "`actual`," which I then blindly made worse in cl/247013322.

RELNOTES=n/a
PiperOrigin-RevId: 770181371