Tweak `ThrowableSubject` null failures to use a `Fact` instance instead of a message.
The exception message looks exactly the same to users either way; this CL is just about how the data is stored in the exception.

In cl/769789387, I used a message because that used to be the only way to do it. But because that CL put special logic into `FailureMetadata` for this case, we can make the case do whatever we want.

`Fact` instances are normally how Truth builds its failure messages, with messages usually just a way for users to supplement those messages as needed. So I think it makes sense to use `Fact` here, as usual. That also lets us use the usual `TruthFailureSubject` for our testing.

RELNOTES=n/a
PiperOrigin-RevId: 770198363