Store all `typeDescriptionOverride` values in `Subject`.
This probably makes the most sense, given that we want a given description to apply only to an exact class (e.g., `ThrowableSubject`), not to a class and its subtypes (e.g., `ThrowableSubject` + `TruthFailureSubject`). We have been able to work that out with constructor parameters, but it requires a lot of boilerplate and a bit of care.

(Either approach is better than trying to use an overridable `String typeDescriptionOverride()` method, which would make it impossible for a `Subject` subclass to set a type description without forcing it onto any subclasses that don't override it—including subclasses outside the package, which _couldn't_ override it unless we made it visible.)

I confirmed that the simplifications in `*StreamSubject` are safe: As the new test in `IntStreamSubjectTest` shows (as does the existing `StreamSubject.testHasSize_fails()`), the failure message still mentions "stream," even though we delegate to an `IterableSubject`. My work to explicitly set a description of "stream" in cl/598964798 and then cl/764978392 looks to have been superstition. (I think the change _might_ have been useful _before_ cl/598964798, but I'm not even sure about that.)

(One thing that I should maybe someday look into: Do we ever want a `typeDescriptionOverride` for a private `Subject` class, such as `class IterableEntries extends IterableSubject` in `MultimapSubject`? My expectation there would be "no" because we end up taking the description "multimap" from `MultimapSubject`, similar to the `StreamSubject` case discussed just above.)

(Bonus: Deal with a few more [occurrences of "subject" where we mean "actual."](https://github.com/google/truth/issues/203)

RELNOTES=n/a
PiperOrigin-RevId: 770384151