Handle [null values](https://github.com/google/truth/issues/207) in `BigDecimalSubject`.
In this case, I'm thinking that nulls are actually fine? It seems reasonable for `isEqualToIgnoringScale` to be strictly less picky than `isEqualTo`. I don't see discussion in cl/91722692, so it may just have never come up. We then added `checkNotNull` to preserve behavior when we adopted nullness checking.

RELNOTES=n/a
PiperOrigin-RevId: 771056906