Consistently use `actualCustomStringRepresentationForPackageMembersToCall()`, standardizing on helper methods.
We currently use a mix of it and `actual`, the latter of which doesn't respect any custom formatting that the type has defined in an `actualCustomStringRepresentation()` override. (I think we use `actual` only because we forget about custom string representations, perhaps motivated by how long "`actualCustomStringRepresentationForPackageMembersToCall()`" is.)

We already had some helper methods, but now we have more.

RELNOTES=n/a
PiperOrigin-RevId: 771287616