Standardize on past tense.
(It took me surprisingly long to notice that "could not" and "is" do not sit well together :))

The past tense is what Truth normally uses, as in "expected," "but was," and "multimap was."

(I notice that we do still use phrasing like "_cannot_ perform assertions on the contents of a null multiset." For some reason, that doesn't really bother me, and I don't think it would work as well as "_could not_ perform...." I guess it's phrased as more of a universal statement, rather than as a statement about this particular assertion and actual value? Still, I grant that we might not have a good reason for mixing the two styles. I'd have to investigate more.)

RELNOTES=n/a
PiperOrigin-RevId: 773354920