[Fail instead of throwing](https://github.com/google/truth/issues/207) for null maps and multimaps (and negative sizes).
This CL includes a larger change to the behavior of at least one edge-case assertion:

```
assertThat(nullMap).containsAtLeastEntriesIn(empty);
```

Prior to this CL, that assertion would succeed. Now, it will fail.

Additionally, this CL fixes yet another parameter name that I missed. This time, it's the one for `MultimapSubject.isEqualTo`.

As a minor implementation point, this CL renames "`getCastSubject`" to "`castActual`," eliminating what I hope is one of the final usages of "subject" to refer to the value under test. This is just in implementation code, though, not in Javadoc or public parameter names.

RELNOTES=Changed assertions like `assertThat(nullMap).isEmpty()` to fail with a useful failure message instead of throwing `NullPointerException`, and similarly for other "bogus" values, such as negative sizes.
PiperOrigin-RevId: 773487350