Introduce a method for retrieving the name of a class.
For now, this is a no-op.

The next step will be to provide implementations that do something other than just call `getName()`. This is not particularly useful on the JVM, but it can be useful in environments that don't have a useful name for all types. (It may even be useful on the JVM, since we can pick a friendlier output format where possible, such as using the canonical name instead of the binary name or omitting "`java.lang.`")

RELNOTES=n/a
PiperOrigin-RevId: 773872316