Modernize and improve `TruthAssertThatTest.java`.
The old code had a method name that claimed to produce a `TypeToken` for a method return type but actually returned it for the single parameter of the input method. Anyway, we can check both the return type and the parameter type, and we can do that just by converting those generic types to strings. Those strings aren't exactly specified, but the same type will certainly produce the same string, and in practice the strings are unambiguous. (Two different types won't have the same `Type.toString()`.) So no need for `TypeToken`.

RELNOTES=n/a
PiperOrigin-RevId: 797792994