Move the proto extension `package-info` to `liteproto`
These extensions both share the same package, but only full has the `package-info`. This means that users of `liteproto` can have an inconsistent view of the package based on whether they also depend on full.

Thankfully full already depends upon lite so we don't need to restructure the build setup.

RELNOTES=n/a
PiperOrigin-RevId: 807453120